Split available variables to its own services
