+++ src/Services/AvailableVariables/BlueprintVariableMapper.php
@@ -1,0 +1,256 @@
+<?php
+
+namespace Justbetter\StatamicStructuredData\Services\AvailableVariables;
+
+use Illuminate\Support\Collection as SupportCollection;
+use Statamic\Facades\Collection as CollectionFacade;
+use Statamic\Facades\Fieldset as FieldsetFacade;
+use Statamic\Fields\Blueprint;
+
+class BlueprintVariableMapper
+{
+    public function fieldTypeIsEligible(string $fieldType): bool
+    {
+        $eligibleFieldTypes = [
+            'text',
+            'assets',
+            'bard',
+            'toggle',
+            'select',
+            'integer',
+            'slug',
+            'date',
+            'entries',
+            'aardvark_seo_meta_title',
+            'aardvark_seo_meta_description',
+        ];
+
+        return in_array($fieldType, $eligibleFieldTypes, true);
+    }
+
+    /**
+     * @param  mixed  $items
+     * @return array<int, array<string, mixed>>
+     */
+    public function normalizeBlueprintItems($items): array
+    {
+        if ($items instanceof SupportCollection) {
+            $items = $items->all();
+        }
+
+        if (! is_array($items)) {
+            return [];
+        }
+
+        $normalizedItems = [];
+
+        foreach ($items as $item) {
+            if ($item instanceof SupportCollection) {
+                $item = $item->all();
+            }
+
+            if (! is_array($item)) {
+                continue;
+            }
+
+            $import = $item['import'] ?? null;
+            if (is_string($import) && $import !== '') {
+                $normalizedItems = array_merge($normalizedItems, $this->getImportedFieldsetItems($item));
+
+                continue;
+            }
+
+            if (isset($item['handle']) && is_array($item['field'] ?? null)) {
+                /** @var array<string, mixed> $item */
+                $normalizedItems[] = $item;
+            }
+
+            $nestedItems = $item['fields'] ?? null;
+            if (is_array($nestedItems) || $nestedItems instanceof SupportCollection) {
+                $normalizedItems = array_merge(
+                    $normalizedItems,
+                    $this->normalizeBlueprintItems($nestedItems)
+                );
+            }
+        }
+
+        return $normalizedItems;
+    }
+
+    /**
+     * @param  array<string, mixed>  $importConfig
+     * @return array<int, array<string, mixed>>
+     */
+    public function getImportedFieldsetItems(array $importConfig): array
+    {
+        $fieldsetHandle = $importConfig['import'] ?? null;
+        if (! is_string($fieldsetHandle) || $fieldsetHandle === '') {
+            return [];
+        }
+
+        $prefix = $importConfig['prefix'] ?? null;
+        if (! is_string($prefix)) {
+            $prefix = '';
+        }
+
+        $fieldset = FieldsetFacade::find($fieldsetHandle);
+
+        if (! $fieldset) {
+            return [];
+        }
+
+        $items = $this->normalizeBlueprintItems($fieldset->fields()->items());
+
+        if ($prefix === '') {
+            return $items;
+        }
+
+        return array_map(function (array $item) use ($prefix): array {
+            $handle = $item['handle'] ?? null;
+
+            if (is_string($handle) && $handle !== '') {
+                $item['handle'] = $prefix.$handle;
+            }
+
+            return $item;
+        }, $items);
+    }
+
+    /**
+     * @param  array<string, mixed>  $field
+     * @return array{name: string, description: string, children: array<int, array<string, mixed>>}|null
+     */
+    public function setFieldData(array $field, ?string $name = null, ?string $description = null, bool $recursive = true): ?array
+    {
+        $fieldHandle = $field['handle'] ?? null;
+        /** @var array<string, mixed> $fieldConfig */
+        $fieldConfig = is_array($field['field'] ?? null) ? $field['field'] : [];
+        $fieldType = $fieldConfig['type'] ?? null;
+
+        if (! is_string($fieldHandle) || $fieldHandle === 'parent' || ! is_string($fieldType)) {
+            return null;
+        }
+
+        $isGroupField = $fieldType === 'group';
+        if (! $isGroupField && ! $this->fieldTypeIsEligible($fieldType)) {
+            return null;
+        }
+
+        /** @var array<int, array<string, mixed>> $children */
+        $children = [];
+
+        if ($fieldType === 'entries' && $recursive) {
+            $collections = is_array($fieldConfig['collections'] ?? null) ? $fieldConfig['collections'] : null;
+
+            if (! isset($collections[0]) || ! is_string($collections[0])) {
+                return null;
+            }
+
+            $children = $this->getCollectionVariables($collections[0], $field);
+        }
+
+        if ($isGroupField && $recursive) {
+            $children = $this->getGroupVariables($field, $fieldConfig);
+        }
+
+        $descriptionValue = is_string($description) ? $description : null;
+        $display = is_string($fieldConfig['display'] ?? null) ? $fieldConfig['display'] : null;
+
+        if ($isGroupField && empty($children)) {
+            return null;
+        }
+
+        return [
+            'name' => $name ?? $fieldHandle,
+            'description' => $descriptionValue ?? $display ?? $fieldHandle,
+            'children' => $children,
+        ];
+    }
+
+    /**
+     * @param  array<string, mixed>  $field
+     * @return array<int, array<string, mixed>>
+     */
+    public function getCollectionVariables(string $collectionHandle, array $field): array
+    {
+        if (! $collectionHandle || $collectionHandle === 'structured_data_templates') {
+            return [];
+        }
+
+        $collection = CollectionFacade::find($collectionHandle);
+
+        if (! $collection) {
+            return [];
+        }
+
+        $blueprint = $collection->entryBlueprints()->first();
+
+        if (! $blueprint) {
+            return [];
+        }
+
+        return $blueprint->fields()->items()->map(function (array $entryField) use ($field) {
+            $entryFieldConfig = is_array($entryField['field'] ?? null) ? $entryField['field'] : [];
+            $parentFieldConfig = is_array($field['field'] ?? null) ? $field['field'] : [];
+            $fieldHandle = is_string($field['handle'] ?? null) ? $field['handle'] : '';
+            $entryFieldHandle = is_string($entryField['handle'] ?? null) ? $entryField['handle'] : '';
+            $name = $fieldHandle.':'.$entryFieldHandle;
+            $parentDisplay = is_string($parentFieldConfig['display'] ?? null) ? $parentFieldConfig['display'] : '';
+            $entryDisplay = is_string($entryFieldConfig['display'] ?? null) ? $entryFieldConfig['display'] : '';
+            $description = $parentDisplay.': '.($entryDisplay ?: $entryFieldHandle);
+
+            return $this->setFieldData($entryField, $name, $description, false);
+        })->filter()->values()->all();
+    }
+
+    /**
+     * @param  array<string, mixed>  $field
+     * @param  array<string, mixed>  $fieldConfig
+     * @return array<int, array<string, mixed>>
+     */
+    public function getGroupVariables(array $field, array $fieldConfig): array
+    {
+        $groupFields = $this->normalizeBlueprintItems($fieldConfig['fields'] ?? []);
+        $groupHandle = is_string($field['handle'] ?? null) ? $field['handle'] : '';
+        $groupDisplay = is_string($fieldConfig['display'] ?? null) ? $fieldConfig['display'] : $groupHandle;
+
+        return collect($groupFields)
+            ->map(function (array $groupField) use ($groupHandle, $groupDisplay): ?array {
+                /** @var string $childHandle */
+                $childHandle = $groupField['handle'];
+
+                $childConfig = is_array($groupField['field'] ?? null) ? $groupField['field'] : [];
+                $childDisplay = is_string($childConfig['display'] ?? null) ? $childConfig['display'] : $childHandle;
+
+                return $this->setFieldData(
+                    $groupField,
+                    $groupHandle.':'.$childHandle,
+                    $groupDisplay.': '.$childDisplay,
+                    false
+                );
+            })
+            ->filter()
+            ->values()
+            ->all();
+    }
+
+    /**
+     * @param  iterable<int, Blueprint>  $blueprints
+     * @return array<int, array<string, mixed>>
+     */
+    public function mapBlueprintsToVariables(iterable $blueprints): array
+    {
+        /** @var array<int, array<string, mixed>> $fieldsArray */
+        $fieldsArray = collect($blueprints)->filter()->reduce(function (array $carry, $blueprint): array {
+            $items = $this->normalizeBlueprintItems($blueprint->fields()->items());
+
+            return array_merge($carry, $items);
+        }, []);
+
+        return collect($fieldsArray)
+            ->map(fn (array $field) => $this->setFieldData($field))
+            ->filter()
+            ->values()
+            ->all();
+    }
+}
