+++ src/Services/AvailableVariables/VariableProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Justbetter\StatamicStructuredData\Services\AvailableVariables;
+
+interface VariableProvider
+{
+    /**
+     * @return array<int, mixed>
+     */
+    public function variables(mixed $parent = null): array;
+}
